+++ src/binder/test-helpers.ts
@@ -4,7 +4,7 @@ import type { ResolvedType } from '../resolver/type-resolution';
 import { BindingContext, StringEncoding } from '../resolver/types';
 import { createMemoryStorer } from './to-abi';
 import { createMemoryLoader } from './to-js';
-import type { JsValue } from '../marshal/types';
+import type { JsValue } from '../marshal/model/types';
 
 export function storeToMemory(ctx: BindingContext, ptr: number, type: ResolvedType, jsValue: JsValue, stringEncoding: StringEncoding, canonicalResourceIds: Map<number, number>): void {
     createMemoryStorer(type, stringEncoding, canonicalResourceIds)(ctx, ptr, jsValue);
